// src/app/Models/developer.model.ts
export interface Developer {
    name: string;
    projects: string;
    unit: string;
    unitType: string;
    level: string;
    location: string;
    exposure: string;
    size: number;
    bedRoom: number;
    bathRoom: number;
    parking: boolean;
    locker: boolean;
    balcony: boolean;
  }